Added a que system.
Since the quotes are filtered inorder ensure short length and avoid repeating the same quotes, the requests can take a few seconds. This is too long for the user to wait every time. So we que up 100 quotes and fill the que when the que is below a certain threshold. Thus the quotes are refreshed instantly.